refactor(intelligence,http): split management/ai_gen.rs by topic, move AI subset (#656) (#658)
The original `management/ai_gen.rs` (~994L) was a misnamed dumping
ground. Only ~25% of it was actually AI-generation; the rest was
chaos config, network profiles, recorder-driven OpenAPI inference,
and rule-explanation storage that happened to share a file.

## Phase A — split by topic (no semantic change)

Four files under `management/`:

- `ai_gen.rs` — kept the `generate_ai_spec` handler + extraction helpers,
  now ~225L (pre-Phase-B). Pure AI, no state usage.
- `traffic_to_openapi.rs` (new) — `generate_openapi_from_traffic`.
  Uses `mockforge-recorder`, can't move to intelligence (recorder→
  intelligence cycle), stays in http.
- `rule_explanations.rs` (new) — list/get/learn handlers that share
  `ManagementState.rule_explanations`. Tightly coupled to state,
  stays in http.
- `chaos_admin.rs` (new) — chaos config + network profiles. Never
  AI-related; chaos-coupled handlers are out of scope for the #555 /
  #656 drain.

`management/mod.rs` glob-re-exports from all four so route wiring is
unchanged. `traffic_to_openapi.rs` cfg-gates its axum imports to match
the `behavioral-cloning`-gated handler (otherwise unused-imports trips
clippy on default-feature builds).

## Phase B — move trimmed ai_gen.rs to mockforge-intelligence

- `mockforge-intelligence/src/handlers/spec_generation.rs` (new) —
  verbatim move of `generate_ai_spec` + extraction helpers. Dropped
  the `State<ManagementState>` extractor (it was never read — axum
  allows handlers to skip state even on stateful routers).
- `mockforge-http/src/management/ai_gen.rs` — now a 19-line shim
  re-exporting from intelligence.
- `data-faker` feature now propagates from http to intelligence so
  the dual data-faker / stub-503 contract stays on a single toggle.

## Net change

- mockforge-http loses ~990L from `management/ai_gen.rs`, gains
  ~805L spread across three topic files (the difference is dedup'd
  imports), and gains a 19L shim.
- mockforge-intelligence gains ~258L (the moved spec_generation
  handler) + 5L of Cargo.toml feature plumbing.
- Total mockforge-http drain attributable to #656 (combined with
  #657 ai_handler move): ~530L of AI code out.

## Verified

- cargo build -p mockforge-intelligence -p mockforge-http
- cargo clippy -p mockforge-intelligence -p mockforge-http --all-targets -- -D warnings
- cargo clippy -p mockforge-intelligence -p mockforge-http --all-targets --features mockforge-http/data-faker -- -D warnings
- cargo test -p mockforge-intelligence (343/343 pass)
- cargo test -p mockforge-http (571/571 pass)
- cargo fmt --all --check

Refs #656. Stacks on #657 (ai_handler move). All AI code that can move
cleanly is now in mockforge-intelligence; the rest stays in http for
documented dep-cycle / state-coupling reasons.
